feat(project config): move dataclass to datamodel package
